Rename rl_agent and Agent directories to agents
