+++ src/agents/SavedPolicy.cc
@@ -1,4 +1,4 @@
-#include <rl_agent/SavedPolicy.hh>
+#include <agents/SavedPolicy.hh>
 #include <algorithm>
 
 SavedPolicy::SavedPolicy(int numactions, const char* filename):
